BackEnd/requirements.txt
fastapi
uvicorn
python-dotenv
motor
pymongo
google-generativeai
google-genai 
faster-whisper
celery
redis
python-multipart
boto3
passlib[bcrypt]
python-jose
email-validator